Included the config.hxx file in log4CPlus source tree.
Included the config.hxx file in log4CPlus source tree.  

This file is normally created at compile time by the macros defined in Log4Cplus. These macros are executed when one builds log4Cplus. The resulting config.hxx is moved to the cmake build snapshot directory and the associated path is added to the include directives — unfortunately the include directives are correctly specified only when building Log4Cplus (other projects do not run the config.hxx code generating Macro ). SImply copying the file to the Log4CPlus source tree is a workaround. Alternatively (better option), the log4CPlus macro should also copy the file in the original log4CPlus source tree or the exact path of the generated config.hxx in the build directory should be referenced in the include directives.